Move scope map to data structures
