+++ conder_core/src/main/data_structures/scope_map.ts
